Borrow groupedEach from Discourse
Using groupedEach seems to have marked performance improvements on my
Nexus 4. To this end I've also marked the recipe names as unbound in the
category template.

To comply with Discourse's GPL 2 licensing this project is now licensed
under GPL 2. I've done my best to include everything necessary to comply with
that license.